#2144 add errorBoundary visual components to storybook log error stack to Sentry too define errorMessages for Panel and Layout components separately
