+++ app/components/eto/dashboard/submitProposalWidget/SubmitProposalWidget.tsx
@@ -7,8 +7,8 @@ import { compose } from "redux";
 import { actions } from "../../../../modules/actions";
 import { appConnect } from "../../../../store";
 import { ButtonArrowRight } from "../../../shared/buttons";
-import { createErrorBoundary } from "../../../shared/ErrorBoundary";
-import { ErrorBoundaryPanel } from "../../../shared/ErrorBoundaryPanel";
+import { createErrorBoundary } from "../../../shared/errorBoundary/ErrorBoundary";
+import { ErrorBoundaryPanel } from "../../../shared/errorBoundary/ErrorBoundaryPanel";
 import { Panel } from "../../../shared/Panel";
 
 import * as styles from "../../etoContentWidget.module.scss";
