+++ app/components/eto/registration/Start.tsx
@@ -5,8 +5,8 @@ import { actions } from "../../../modules/actions";
 import { appConnect } from "../../../store";
 import { onEnterAction } from "../../../utils/OnEnterAction";
 import { LayoutAuthorized } from "../../layouts/LayoutAuthorized";
-import { createErrorBoundary } from "../../shared/ErrorBoundary";
-import { ErrorBoundaryLayoutAuthorized } from "../../shared/ErrorBoundaryLayoutAuthorized";
+import { createErrorBoundary } from "../../shared/errorBoundary/ErrorBoundary";
+import { ErrorBoundaryLayoutAuthorized } from "../../shared/errorBoundary/ErrorBoundaryLayoutAuthorized";
 import { LoadingIndicator } from "../../shared/loading-indicator";
 import { EtoRegistrationPanel } from "./EtoRegistrationPanel";
