+++ app/components/landing/Landing.tsx
@@ -7,8 +7,8 @@ import { compose } from "redux";
 import { etoCompaniesCards } from "../../data/etoCompanies";
 import { withContainer } from "../../utils/withContainer";
 import { LayoutUnauthorized } from "../layouts/LayoutUnauthorized";
-import { createErrorBoundary } from "../shared/ErrorBoundary";
-import { ErrorBoundaryLayoutUnauthorized } from "../shared/ErrorBoundaryLayoutUnauthorized";
+import { createErrorBoundary } from "../shared/errorBoundary/ErrorBoundary";
+import { ErrorBoundaryLayoutUnauthorized } from "../shared/errorBoundary/ErrorBoundaryLayoutUnauthorized";
 import { EtoCard } from "../shared/EtoCard";
 import { Features } from "./parts/Features";
 import { LandingFeatured } from "./parts/LandingFeatured";
