--- app/components/shared/ErrorBoundary.module.scss
@@ -1,7 +1,0 @@
-.panel {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
