+++ app/components/shared/errorBoundary/ErrorBoundaryLayoutUnauthorized.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+import { FormattedMessage } from "react-intl-phraseapp";
+
+import { LayoutUnauthorized } from "../../layouts/LayoutUnauthorized";
+
+import * as styles from "./ErrorBoundary.module.scss";
+
+export const ErrorBoundaryLayoutUnauthorized: React.SFC<{}> = () => {
+  return (
+    <LayoutUnauthorized>
+      <div className={styles.errorLayout}>
+        <FormattedMessage id="error-boundary.main-error-message" />
+      </div>
+    </LayoutUnauthorized>
+  );
+};
