+++ app/components/shared/errorBoundary/ErrorBoundaryPanel.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { FormattedMessage } from "react-intl-phraseapp";
 
-import { PanelBase } from "./PanelBase";
+import { PanelBase } from "../PanelBase";
 
 import * as styles from "./ErrorBoundary.module.scss";
