+++ app/components/wallet-selector/WalletSelector.tsx
@@ -19,8 +19,8 @@ import { appRoutes } from "../appRoutes";
 import { LayoutRegisterLogin } from "../layouts/LayoutRegisterLogin";
 import { LayoutUnauthorized } from "../layouts/LayoutUnauthorized";
 import { Button, ButtonLink, EButtonLayout } from "../shared/buttons";
-import { createErrorBoundary } from "../shared/ErrorBoundary";
-import { ErrorBoundaryLayoutUnauthorized } from "../shared/ErrorBoundaryLayoutUnauthorized";
+import { createErrorBoundary } from "../shared/errorBoundary/ErrorBoundary";
+import { ErrorBoundaryLayoutUnauthorized } from "../shared/errorBoundary/ErrorBoundaryLayoutUnauthorized";
 import { ICBMWalletHelpTextModal } from "./ICBMWalletHelpTextModal";
 import { WalletMessageSigner } from "./WalletMessageSigner";
 import { WalletRouter } from "./WalletRouter";
